Version 0.5.0-BETA
+New GUI powered by JavaFX with more extensible functionality
+Updated transaction system to fully support new address structure
+Updated addresses to increase security and anonymity

-Removed old Swing powered GUI. Code still there for historical reasons, just not hooked up

=Fixed network bugs that were hanging the network process when sending complex packets (new transaction packets with new address structure)
=Fixed bug with block payouts, specifically genesis
=Fixed bug with transaction management system not verifying new address structures correctly in certain cases

Testing that has been completed:
-GUI elements have been thoroughly tested
-GUI backend has light testing

Testing that has yet to be completed:
-Functions that previously worked on the old GUI, which appear to work on the new GUI, but are not fully verified to be completely functional yet.

This commit is mostly about the new GUI system, which, until further notice, is licensed under a full copyright with all rights reserved. The GUI will most likely take on the same license as the rest of the program, but is licensed this way currently to help prevent complete copy of the program but allow for other developers to use the core still.

Various stabilizations were added to the program that are too numerous to mention here, most have to do with the networking and transaction systems, which have had recent issues with the new address system.

Next update will include finalization of the GUI, and relay->relay communication protocols.