Version 0.5.1-BETA
+Added basis for multi-device mining ability once mining is released publicly (GPU mining as well as CPU mining)
+Added secondary window for the GUI that will give verbose details on transaction history which is accessible by clicking on a recent transaction
+Added new labelling and structuring to the new GUI to more aptly describe the functions of various elements
+Added basis for relay->relay communication protocols
+Added 100 extra tokens for future use with new assets we want to add to the currency, they will remain dormant until that time
+Added ability to choose how many cores of a CPU are dedicated to mining when CPU mining

=Finalized GUI and fixed multiple bugs with it, the most prominent being incorrectly showing balance if the balance for certain tokens was zero
=Restructured parts of the codebase to make more logical and logistical sense, in particular moving parts of the mining program out of the chain management system and into their own system
=Finalized the block chain. The system has been thoroughly tested and used and is ready to be declared final. The chain that currently exists will be the chain the program releases on.

-Removed small pieces of non-functional or obselete code to help start cleaning the codebase to ready it for it's first public beta version.

Testing that has been completed:
-All GUI funcationality is tested and working
-Blockchain system with new tokens
-Transaction system with new tokens

Testing that has yet to be completed:
-None, once the relay->relay system is actually up, this will be the next item to test

This commit finalizes quite a few key parts of the program and starts moving the program towards the public beta range. There will be a feature freeze in the near future to focus on stability and streamlining before the program is publicly released.

There were not as many stabilizations added this version as there were in previous ones, as the program is beginning to mature to the point that it does not need as many. There will be a huge push for rock solid stability and reliability after the feature freeze occurs.

Next update will have the first relay->relay communication protocols in action, and should see some improvements to the GUI backend, and possibly GPU mining (developer side only)