Version 0.5.2-BETA
+Added Origin logo as taskbar icon
+Added new network code that will send all pending transactions to a node from the relay after a connection is established, this is crucial for mining and for relay-relay communication purposes

=Fixed bug where the "Address" text box would start focused so the promt text would not be seen unless it was clicked out of

Testing that has been completed:
Same as last build, this is more an intermidiate that is here to be the last version before relay-relay comm goes online, so there is a clear working version before the network code begins to be added to

Testing that has yet to be completed:
The new network code is not fully tested yet, but is very simple and should work just fine. This is the last groundwork that needed to be laid for the relay-relay comms

This is a small update with few additions and reworks that will be the last before relay-relay comms are committed.