Version 0.6.0-BETA
+Added relay-relay communications protocols (1st iteration)
+Added new packets to support relay-relay comms

=Updated many network files for relay comms

Testing that has been completed:
Base program with single relay mining and multi-relay connecting

Testing that has yet to be completed:
full relay-relay comms

This may not look like it, but is a massive update that took many hours of research and development to form a suitable protocol that would serve relay-relay communications efficiently and quickly. This is the first iteration and many changes are expected in the upcoming patches. After this feature is finished the program will have all the necessary components to act as a full cryptocurrency. (Before now has been a single-relay cryptocurrency, which is a centralized system).

This is a huge update but nothing will seem new on the client side, all of the major updates are backend only, although there have been slight changes made to standard client node comms to allow for an easier implementation of the relay-relay comms.

There is a stress test scheduled for the currency in the coming days that will give us plenty of information on how the networking system will react to incredibly heavy loads with 50MB block sizes and many transactions from all tokens, the outcome of this test will let us know if the current network system is suitable for the network loads expected when the cryptocurrency is in its full form and released into the wild.