Version 0.7.0-BETA
+Added completely new network system which is more modular and streamlined
+Added iteration 2 of relay-relay comms
+Added multi-relay connections from nodes capability
+Added monoimplementation design for connections
+Added height indicator to the GUI under the wallet tab
+Massive code additions because of new network system

-Removed old gui system
-Removed entire old network system
-Massive code deletions because of new network system

Testing that has been completed:
Connections are up and working, packets seem to be directing and processing correctly, not everything is fully functional yet

Testing that has not been completed:
Relay-Relay comms haven't been fully tested yet, but are expected to work much better than before, as the basis of the rewrite was to make the comms better

This is a MASSIVE update, the ENTIRE network system has been COMPLETELY rewritten. As such, there are expected to be some bus and there will be some turbulence with operation in the coming days. The next release is expected to fix all the breaking bugs of this update and make it stable enough to run as well as it did before relay-relay comms were started.