Version 0.7.1-BETA
+Added iteration 3 of relay-relay communications. Now mitigating collisions successfully!
+Added more modularity to chain management system

=Bug fixes

-Removed small pieces of code here and there

Testing that has been completed:
Collision mitigation of 2 competing chains fully mitigated efficiently within about 20 seconds

Testing that has yet to be completed:
general day to day operation of the program

After a few GUI features and updates are added, the feature freeze will be done and we will be making the final push towards Open Beta. This is the final stretch.