Version 0.7.2-BETA
+Added some extras to the new network system

=Fixed packet duplication bug with network system
=Fixed relays not relaying
=Fixed bad logic on discrepancy detection
=Added extra memory to network system, still much smaller than old system, optimization of this will happen during the feature freeze
=Fixed various packet processing bugs

-Removed some bad code in the new network system

Testing that has been completed:
Mining
Transactions
Chain updating
Collision mitigation
Network heavy loads

Testing that has not yet been completed:
Live test with multiple relays trying to induce a collision in real time instead of a synthetic collision

This is a bug fix update, lots of testing and work was put in to fixing all the bugs with the new network system, there will probably be more, but a great deal are fixed with this update. The issue tracker will be used more from this update to keep pace with new issues as they occur, as the programs complexity now requires this

After a few GUI features are added in the coming updates, we will move to our feature freeze to stabilize the codebase, then we'll be ready for an open beta.