Version 0.9.0
+Completely rewrote entire network system to use Netty, an industrial strength networking library far superior to KryoNet, the previously used library
+Updated all string representations of binary to now be base64, to save space and network bandwidth
+Added new features to the CPU miner, and did serious optimizations on it
+Added new features throughout the program, this commit is so incredibly large that not all of the changes are presently remembered

=Fixed many network related bugs and a few block and miner related bugs

-Removed old networking library and some defunct code

Testing that has been completed:
network entirety with light transaction testing

Testing that has yet to be completed:
Heavy (1000+ transactions) network loading and unloading while mining

This is a massive commit as stated above. The program may not look very different on the outside but almost every file in the program was modified in some way to create this commit. The reliability of the network in the application is much higher now, and sending packets is also much faster.