Version 0.10.0-BETA
This commit is all about optimization

+Added new library to handle database for UTXOs, by our measurements, it's close to 1,300,000% faster than the previously used library
+Added more concurrency to block and transaction verification to speed things up a bit (280% speed gain)
+Added some more new network code to make sending/receiving/processing packets faster (120% speed gain)

=Fixed many bugs related to transactions when large quantities of transactions were sent out

-Removed some bloat throughout the program, although I do not have an exact number for the gains in performance from this, I'd wager them to be between 10-40% generally depending on which mechanisms they belogned to

Testing that has been performed:
Heavy transaction usage in blocks, verification of transactions and blocks, mining with multiple peers connected, chain integrity, network integrity, transaction integrity

Testing that has yet to be completed:
First large scale test

This is the beginning of the end. This program now meets or exceeds specifications laid out to accomplish the task it was given. Soon we will release the program into a semi-open beta to start getting some large scale testing going. The last (important) part that needs to be taken care of is networking for new relayers who want to join the network. This won't be very hard and can be set up quickly.