Version 0.13.6-BETA
=Fixed memory leak from GUI backend,(for real this time) should hopefully fix longevity issues (around 6 days after program is started it would sometimes crash)
+Optimizations to some GUI code

Testing that has been completed:
Ran program (while mining for worst case scenario) for 2+ hours and did not experience the same memory oscillation indicative of the leak.

Testing that has not yet been completed:
Running program for a few days and watching memory usage

Small release but this is a very important bugfix, this brings the program one step closer to being production ready, as it is a huge stability update.