Version 0.16.0-BETA
+New state management system that is faster, more stable, and more reliable
=Major bug fixed that allowed possible double spend, as shown by the cain issue at 14536

Testing that has been completed:
Collision mitigation, double spending, chain management, network management.

Testing that has yet to be completed:
None that we know of

This is a huge release. Although you won't notice anything different in the front end, a large part of the backend has been overhauled. As this is such a huge change, there may be some bugs that we've missed, but we did a large amount of stress testing and we believe it should be far more stable than the old versions.

As mentioned in an announcement, we'll be rolling the chain back to 14356 because a double spend made it onto the block chain. The new version has a very specific fix to fix this issue, and this will not happen again.