Version 0.16.1-BETA
-Removed some old useless network code
=Fixed many bugs with network code
=Fixed bugs in state manager code
=Fixed bugs in main runninng code
=Fixed bugs with collision mitigation
=Fixed bugs with lite client GUI
=Fixed bugs with full node block storage system
=Fixed many other bugs

Testing that has been completed:
Everything we can think of. Relay->Relay is still a bit shaky, but will be stable by next release

Testing that has not yet been completed:
None that we know of.

This is a huge bug fix release, please report any issues to the bitbucket or discord.