Version 0.18.0-BETA
This is the testing release for 0.18.0, with the main test point being the ADX right now, along with some other bugs with the NovusTX system.

Here is a relatively simplified changelog, since so much has changed:

+ASE (Ampex Scripting Engine) for powering P2SH transactions
+ADX (Ampex Decentralized Exchange) which runs on the P2SH system for an exchange on the block chain
+NovusTX the new transaction system, completely revamped from the old one for speed and data savings, this system is the groundwork for everything else
+New key support for ED25519 keys
+Multi-length address and address prefixing support
+Lots of other goodies from the P2SH system (multi-sig, micro-payment channels, etc)
+additions to nearly every system
+Larger UI, with more features and better scaling for the most part
+there's literally too many changes to put here
+Colors in output in console for easy debugging

=fixed PPLNS bug on pools

-Old transaction system is functional, but no longer used for transactions, meaning old transactions will verify, but new ones cannot be created on the old system
-Removed a lot of defunct code

Testing that has been completed:
Tons, nearly every system has been tested. The ADX is the final system that is undergoing beta testing before release on the main chain currently

Testing that has not yet been completed:
Some P2SH stuff still needs testing, will be tested after the ADX is done

This is the single largest release we've ever pushed or worked on, version 0.18.0 is going to be incredibly massive once it is released publicly. This also means there will be plenty of new bugs in it. Version 0.19 will focus on fixing these, and very few features will be implemented between now and version 1.0.0, which is scheduled after 0.19