v 0.18.0
=Fixed chain conversion tools

This is the final build of 0.18.0. There is a large list of changes that have occurred since 17.7, but I'm not going to list them all here, the previous commit messages will show most of those changes.

The big ones are:
-New address system
-New transaction system
-P2SH system
-Decentralized exchange (not enabled presently)

and a few other minor features. This is a massive release, and, even though we've tested everything, there are expected to be a few bugs.

If you're running a full node on version 17.7, you'll have to delete your chain and redownload. 17.5 and 17.6 chains should be converted over to this system by the program, but if there's a problem, delete them and redownload.

Testing that has been completed:
Everything, should be a very stable release
Testing that has not yet been completed:
Nothing